added some prints

+++ src/modradrrtmgyuriroutines.f90
@@ -140,7 +140,8 @@ contains
 
 		  do k = 1, n_quantiles
 			 index_q = (float(k) / float(n_quantiles + 1)) * n_s
-			 q(k)    = x(index_q)
+			 ! q(k)    = x(index_q)
+			 q(k)    = x(nint(index_q))
 		  end do
 
 	   end if
